Prevent tasks from being re-run when postInstall() is called twice
It appears that when creating a new project, Composer fires both post-create-project-cmd and post-install-cmd, resulting in postInstall() being run twice. Removing the post-install-cmd trigger COULD fix some issues and simplify things, but I'd like to be able to run `composer install` on the server whenever `composer.lock` is updated, and to have Bootstrap files automatically copied to webroot when that happens.